Update experience.tsx
Added logic to animate VerticalTimeline element. Original code doesnot render the element. Adding visible={true} removes the animation and visible={inView} causes the animation to repeat each time element is in view. If we add triggerOnce: true in useInView, it removes the animation from the header as well. Thus we use another state variable hasBeenViewed to keep a track if whether the element has been viewed.